refactor: enhance error handling and logging in export and view functions
